+++ src/components/spinner.js
