+++ src/containers/router.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
-import App from '../containers/app';
+import App from '../components/app';
 
 export default (
   <BrowserRouter>
